+++ pulsarDb/TimingModel.h
@@ -24,10 +24,10 @@ namespace pulsarDb {
           \param ev_time Time of the event.
       */
       virtual FrequencyEph calcEphemeris(const PulsarEph & eph, const AbsoluteTime & ev_time) const {
-        double dt = (ev_time - eph.epoch()).sec();
-        double f0 = eph.f0() + eph.f1() * dt + eph.f2()/2.0 * dt * dt;
-        double f1 = eph.f1() + eph.f2() * dt;
-        double phi0 = calcPulsePhase(eph, ev_time);
+        long double dt = (ev_time - eph.epoch()).sec();
+        long double f0 = eph.f0() + eph.f1() * dt + eph.f2()/2.0 * dt * dt;
+        long double f1 = eph.f1() + eph.f2() * dt;
+        long double phi0 = calcPulsePhase(eph, ev_time);
         return FrequencyEph(eph.valid_since(), eph.valid_until(), eph.epoch(), phi0, f0, f1, eph.f2());
       }
 
@@ -37,8 +37,8 @@ namespace pulsarDb {
           \param ev_time Time of the event.
       */
       virtual void cancelPdot(const PulsarEph & eph, AbsoluteTime & ev_time) const {
-        double dt = (ev_time - eph.epoch()).sec();
-        double dt_squared = dt * dt;
+        long double dt = (ev_time - eph.epoch()).sec();
+        long double dt_squared = dt * dt;
         ev_time += Duration(eph.f1()/eph.f0()/2.0 * dt_squared + eph.f2()/eph.f0()/6.0 * dt * dt_squared, UnitSec);
       }
 
@@ -47,11 +47,11 @@ namespace pulsarDb {
           \param eph The ephemeris.
           \param ev_time Time of the event.
       */
-      virtual double calcPulsePhase(const PulsarEph & eph, const AbsoluteTime & ev_time) const {
-        double dt = (ev_time - eph.epoch()).sec();
-        double int_part; // ignored, needed for modf.
-        double dt_squared = dt * dt;
-        double phase = modf(eph.phi0() + eph.f0() * dt + eph.f1()/2.0 * dt_squared + eph.f2()/6.0 * dt * dt_squared, &int_part);
+      virtual long double calcPulsePhase(const PulsarEph & eph, const AbsoluteTime & ev_time) const {
+        long double dt = (ev_time - eph.epoch()).sec();
+        long double int_part; // ignored, needed for modf.
+        long double dt_squared = dt * dt;
+        long double phase = std::modf(eph.phi0() + eph.f0() * dt + eph.f1()/2.0 * dt_squared + eph.f2()/6.0 * dt * dt_squared, &int_part);
         if (phase < 0.) ++phase;
         return phase;
       }
@@ -60,7 +60,7 @@ namespace pulsarDb {
           \param eph The ephemeris.
           \param ev_time Time of the event.
       */
-      virtual double calcOrbitalPhase(const OrbitalEph & eph, const AbsoluteTime & ev_time) const;
+      virtual long double calcOrbitalPhase(const OrbitalEph & eph, const AbsoluteTime & ev_time) const;
 
       /** \brief
           \param eph The ephemeris.
