+++ src/TimingModel.cxx
@@ -6,6 +6,8 @@
 #include <cmath>
 #include <stdexcept>
 
+#include <iostream>
+
 #include "pulsarDb/CanonicalTime.h"
 #include "pulsarDb/OrbitalEph.h"
 #include "pulsarDb/PulsarEph.h"
@@ -29,15 +31,15 @@ namespace {
  * solve Kepler equation (KEPLER)  g + e sin E = E
  */
 int atKepler(
-        double g,        /* input: mean anomaly */
-        double eccent,        /* input: eccentricity */
-        double *e)        /* output: eccentric anomaly */
+        long double g,        /* input: mean anomaly */
+        long double eccent,        /* input: eccentricity */
+        long double *e)        /* output: eccentric anomaly */
 {
-    static double eps = 1e-6;
+    static long double eps = 1e-6;
     static int imax = 50;
 
     int i;
-    static double error, deltae, d__1;
+    static long double error, deltae, d__1;
 
     *e = g;
     if (g == 0.) return 0;
@@ -45,7 +47,7 @@ int atKepler(
     for (i=0; i<imax; i++) {
         deltae = (g - *e + eccent * sin(*e)) / (1. - eccent * cos(*e));
         *e += deltae;
-        error = (d__1 = deltae / *e, fabs(d__1));
+        error = (d__1 = deltae / *e, std::fabs(d__1));
         if (error < eps) return 0;
     }
     return NOT_CONVERGED;
@@ -55,19 +57,19 @@ int atKepler(
 
 namespace pulsarDb {
 
-  double TimingModel::calcOrbitalPhase(const OrbitalEph & eph, const AbsoluteTime & ev_time) const {
+  long double TimingModel::calcOrbitalPhase(const OrbitalEph & eph, const AbsoluteTime & ev_time) const {
     // compute elapsed time from epoch of periastron in seconds
-    double delta_second = (ev_time - TdbTime(eph[T0])).sec();
+    long double delta_second = (ev_time - TdbTime(eph[T0])).sec();
 
     // Compute the time difference as a fraction of the period.
-    double delta_period = delta_second / eph[PB];
+    long double delta_period = delta_second / eph[PB];
 
     // Compute the complete phase.
-    double phase = delta_period * (1. - delta_period * eph[PBDOT] / 2.0);
+    long double phase = delta_period * (1. - delta_period * eph[PBDOT] / 2.0);
 
     // Express phase as a value between 0. and 1.
-    double int_part;
-    phase = modf(phase, &int_part);
+    long double int_part;
+    phase = std::modf(phase, &int_part);
     if (phase < 0.) ++phase;
     return phase;
   }
@@ -78,7 +80,7 @@ namespace pulsarDb {
 
   void TimingModel::demodulateBinary(const OrbitalEph & eph, AbsoluteTime & ev_time) const {
     static const int s_max_iteration = 1000;
-    const double epsilon = 10.0e-9; // 10 nanoseconds.
+    const long double epsilon = 10.0e-9; // 10 nanoseconds.
 
     // Save target arrival time (ev_time) in orig_time.
     // TODO: Do not use TdbTime explicitly.  Generalize it.
@@ -102,7 +104,7 @@ namespace pulsarDb {
       // Compare time difference between candidate demodulated time
       // (ev_time) and target arrival time (orig_time) with the
       // estimated orbital delay based on the binary model (delay).
-      if (fabs(((orig_time - ev_time) - delay).sec()) < epsilon) break;
+      if (std::fabs(((orig_time - ev_time) - delay).sec()) < epsilon) break;
     }
 
     // Check for non-convergence.
@@ -116,16 +118,16 @@ namespace pulsarDb {
 
   Duration TimingModel::calcOrbitalDelay(const OrbitalEph & eph, const AbsoluteTime & ev_time) const {
     // compute elapsed time from epoch of periastron in seconds
-    double delta_second = (ev_time - TdbTime(eph[T0])).sec();
+    long double delta_second = (ev_time - TdbTime(eph[T0])).sec();
 
     // calculate mean anomaly
-    double delta_period = delta_second / eph[PB];
-    double mean_anomaly = OrbitalEph::s_two_pi * delta_period
+    long double delta_period = delta_second / eph[PB];
+    long double mean_anomaly = OrbitalEph::s_two_pi * delta_period
       * (1. - delta_period * eph[PBDOT] / 2.0);
 
     // solve Kepler's equasion
-    double eccen = eph[ECC] + eph[ECCDOT] * delta_second; // eccentricity
-    double eccen_anomaly = 0.0; // eccentric anomaly
+    long double eccen = eph[ECC] + eph[ECCDOT] * delta_second; // eccentricity
+    long double eccen_anomaly = 0.0; // eccentric anomaly
     int status = atKepler(mean_anomaly, eccen, &eccen_anomaly);
 
     // atKepler not converged
@@ -134,25 +136,25 @@ namespace pulsarDb {
     }
 
     // convert eccentric anomaly to true anomaly
-    double true_anomaly = 2.0 * atan(sqrt((1.0+eccen)/(1.0-eccen))
+    long double true_anomaly = 2.0 * atan(sqrt((1.0+eccen)/(1.0-eccen))
                                      * tan(eccen_anomaly/2.0));
     true_anomaly += OrbitalEph::s_two_pi * floor((eccen_anomaly - true_anomaly)/ OrbitalEph::s_two_pi);
     while ((true_anomaly - eccen_anomaly) > OrbitalEph::s_one_pi) true_anomaly -= OrbitalEph::s_two_pi;
     while ((eccen_anomaly - true_anomaly) > OrbitalEph::s_one_pi) true_anomaly += OrbitalEph::s_two_pi;
 
     // compute periastron longitude
-    double omega = eph[OM]
+    long double omega = eph[OM]
       + eph[OMDOT] * true_anomaly * eph[PB] / OrbitalEph::s_two_pi;
 
     // compute projected semimajor axis
-    double semiax = eph[A1] + eph[XDOT] * delta_second;
+    long double semiax = eph[A1] + eph[XDOT] * delta_second;
 
     // compute time delays due to orbital motion
-    double roemer_frac = sin(omega) * (cos(eccen_anomaly) - eccen)
+    long double roemer_frac = sin(omega) * (cos(eccen_anomaly) - eccen)
       + sqrt(1.0 - eccen*eccen) * cos(omega) * sin(eccen_anomaly);
-    double roemer = semiax * roemer_frac;
-    double einstein = eph[GAMMA] * sin(eccen_anomaly);
-    double shapiro = - 2.0 * eph[SHAPIRO_R]
+    long double roemer = semiax * roemer_frac;
+    long double einstein = eph[GAMMA] * sin(eccen_anomaly);
+    long double shapiro = - 2.0 * eph[SHAPIRO_R]
       * log(1.0 - eccen*cos(eccen_anomaly) - eph[SHAPIRO_S]*roemer_frac);
 
     // return total delay
